Don't Understand
This isn't saving to the proper location. I'm trying to save all of the
commits I have made this past week, and I think I messed it up really
badly.
This was an attempt to restart the game after it was lost which didn't
work out, but I saved it to the wrong spot so none of my recent changes
have committed. I'm doing this all now in bulk to show that I did work.