Restart
Game does not pause when keyPressed; it restarts instead. However, the
shark now looks like it has a long shadow extending to the left and when
the game is restarted, the total number of raindrops in the array all
fall at once (in this case, amount = 100). This would be great for a
lightning round, but this isn't a true restart of the game.