Update README.md
Notes:
I can't seem to figure out the GitHub issue and I've spent way too much time on it to have no progress. It's getting in the way of me uploading other things, like this. I didn't have a lot of time over this break due to family commitments, but I should have planned for a worst case scenario similar to this. I'll continue trying new ways to upload it.